requirements.txt
opencv-python==4.8.0.76
numpy==1.26.4
colorama==0.4.6
pwinput==1.0.0
pycryptodome==3.20.0 